Changed Firm agent constructors and print methods

+++ src/Consumer_Good.hpp
@@ -9,7 +9,7 @@ class Consumer_Firm_Agent;
 class Consumer_Good: public General_Good{
 
     public:
-    Consumer_Good(Consumer_Firm_Agent*  seller_pointer, int seller_price, int seller_quantity);
+    Consumer_Good(Consumer_Firm_Agent*  seller_pointer, float seller_price, int seller_quantity);
 
     // Getters
     Consumer_Firm_Agent* GetSeller() {return pSeller;}
